Errase white lines

+++ calc.py
@@ -3,6 +3,7 @@
 
 import sys
 
+
 def plus(op1, op2):
     """ Function to sum the operands """
     return op1 + op2
@@ -12,12 +13,14 @@ def minus(op1, op2):
     """ Function to substract the operands """
     return op1 - op2
 
+
 def divide(op1, op2):
     """ Function to divide the operands """
     try:
         return op1 / op2
     except:
         sys.exit("Error: Invalid parameters")
+
 
 def multiply(op1, op2):
     """ Function to multiply the operands """
@@ -26,8 +29,8 @@ def multiply(op1, op2):
 
 if __name__ == "__main__":
     try:
-        operando1 = int(sys.argv[1])
-        operando2 = int(sys.argv[3])
+        operando1 = float(sys.argv[1])
+        operando2 = float(sys.argv[3])
     except ValueError:
         sys.exit("Error: Non numerical parameters")
